py_backend/requirements.txt
python-dotenv
google-generativeai
fastapi 
uvicorn 
pydantic 
python-multipart 